fix: handle nested domains for `isSameCozy`
