+++ src/constants/strings.json
@@ -30,8 +30,6 @@
   },
   "errors": {
     "cozyClientMismatch": "Could not validate redirect url, Cozy Client mismatch",
-    "domainMismatch": "Could not validate redirect url, domain mismatch",
-    "protocolMismatch": "Could not validate redirect url, protocol mismatch",
     "clearPersistentIconTable": "Could not clear icon cache. Saved cache might be corrupted.",
     "getPersistentIconTable": "Could not get icon cache.",
     "setPersistentIconTable": "Could not set icon cache. Cache to save might be corrupted.",
